+++ day06/ex00/lib_uart.h
@@ -19,7 +19,7 @@ void    uart_getstr(char* str, char size);
 void    uart_printstr(const char* str);
 void    uart_printnbr_8bits(uint8_t nb);
 void    uart_printnbr_16bits(uint16_t nb);
-void    uart_printnbr_hex_16bits(uint8_t nb);
+void    uart_printnbr_hex_16bits(uint16_t nb);
 void    uart_printnbr_hex_8bits(uint8_t nb);
 
 #endif
